fix(carrier-cooling): clean up SetCarrierCooling — bare response return, split validator tests, add enum-invalid test coverage
