+++ backend/test/Anela.Heblo.Tests/Application/CarrierCooling/SetCarrierCoolingHandlerTests.cs
@@ -2,7 +2,6 @@ using Anela.Heblo.Application.Features.CarrierCooling.UseCases.SetCarrierCooling
 using Anela.Heblo.Domain.Features.Catalog;
 using Anela.Heblo.Domain.Features.Logistics;
 using FluentAssertions;
-using FluentValidation.TestHelper;
 using Moq;
 using Xunit;
 
@@ -48,54 +47,5 @@ public class SetCarrierCoolingHandlerTests
                     s.ModifiedBy == "user-123"),
                 It.IsAny<CancellationToken>()),
             Times.Once);
-    }
-}
-
-public class SetCarrierCoolingValidatorTests
-{
-    private readonly Mock<IShippingMethodCatalog> _catalogMock = new();
-
-    private SetCarrierCoolingValidator CreateValidator() => new(_catalogMock.Object);
-
-    private void SetupCatalog(params (Carriers, DeliveryHandling)[] options)
-    {
-        _catalogMock.Setup(c => c.GetAvailableDeliveryOptions())
-            .Returns(options.ToList().AsReadOnly());
-    }
-
-    [Fact]
-    public void Validator_PassesForAvailableCombo()
-    {
-        SetupCatalog((Carriers.PPL, DeliveryHandling.NaRuky));
-        var validator = CreateValidator();
-        var request = new SetCarrierCoolingRequest
-        {
-            Carrier = Carriers.PPL,
-            DeliveryHandling = DeliveryHandling.NaRuky,
-            Cooling = Cooling.L1,
-            ModifiedBy = "user-123",
-        };
-
-        var result = validator.TestValidate(request);
-
-        result.ShouldNotHaveAnyValidationErrors();
-    }
-
-    [Fact]
-    public void Validator_FailsForUnavailableCarrierHandlingCombo()
-    {
-        SetupCatalog((Carriers.PPL, DeliveryHandling.NaRuky));
-        var validator = CreateValidator();
-        var request = new SetCarrierCoolingRequest
-        {
-            Carrier = Carriers.GLS,
-            DeliveryHandling = DeliveryHandling.NaRuky,
-            Cooling = Cooling.L1,
-            ModifiedBy = "user-123",
-        };
-
-        var result = validator.TestValidate(request);
-
-        result.ShouldHaveAnyValidationError();
     }
 }
